Update patch set 1
Patch Set 1:

I received the next problem:

The box 'yk0/ubuntu-xenial' could not be found or could not be accessed in the remote catalog.

I think we removed the creation of this box a couple months ago.
We should add this box inside setup_vagrant.sh again, otherwise new user may not find them.

Patch-set: 1