Update patch set 1
Patch Set 1:

To deploy kubeadm_onap I use the command: 'vagrant up', is it ok?

I updated to https://gerrit.opnfv.org/gerrit/#/c/51993/ and I don't have the problem with the box. However, I have an infinite loop with this description: 
onap: + r=0
    onap: + '[' 0 == 0 ']'
    onap: + sleep 30
    onap: + plink -ssh -pw vagrant vagrant@master 'cat ~/.kube/config'
    onap: + echo 'y\n'
    onap: ++ wc -l
    onap: ++ grep 'tiller-deploy.*Run'
    onap: ++ kubectl get pods -n kube-system

It've been showing me this log for 20 minutes. Is it ok?

Patch-set: 1