Update patch set 4
Patch Set 4:

I have this error: 
:(
    worker1: Run 'kubectl get nodes' on the master to see this node join the cluster.
An error occurred while executing multiple actions in parallel.
Any errors that occurred are shown below.

An error occurred while executing the action on the 'master'
machine. Please handle this error then try again:

The SSH command responded with a non-zero exit status. Vagrant
assumes that this means the command failed. The output for this command
should be in the log above. Please read the output to determine what
went wrong.

Patch-set: 4